+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Hyperlink to GoCommerce powered by logo
+
 ## [2.15.3] - 2019-07-29
 ### Changed
 - Use vtex.device-detector for device detection.